pylint.

+++ faucet/dp.py
@@ -331,8 +331,8 @@ configuration.
         test_config_condition(not (self.nd_neighbor_timeout < (self.timeout / 2)), (
             'L2 timeout must be > ARP timeout * 2'))
         if self.cache_update_guard_time == 0:
-            self.cache_update_guard_time = min(
-                self.arp_neighbor_timeout, self.nd_neighbor_timeout, self.timeout) / 2
+            self.cache_update_guard_time = int(min(
+                self.arp_neighbor_timeout, self.nd_neighbor_timeout, self.timeout) / 2)
         if self.learn_jitter == 0:
             self.learn_jitter = int(max(math.sqrt(self.timeout) * 3, 1))
         if self.learn_ban_timeout == 0:
